Change Repository structs
The previous Repository struct didn't have all fields returned for the GET operation. These were added.

On both Repository and RepositoryConfig we turn SendBuildStatus and CommentOnPullRequests to pointers, so we are able to differentiate null from false.
Coveralls API docs tell they are optional boolean fields. In practice one defaults to false and the other to null, but really none is required on POST.